requirements.txt
pip==23.1.2
streamlit==1.33.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
ultralytics==8.2.2 
Pillow==9.4.0
ffmpeg-python==0.2.0 
datetime==5.5
fastapi==0.110.2 
uvicorn==0.29.0
python-multipart==0.0.9 
ngrok==1.2.0 
numpy==1.25.2 
matplotlib==3.7.1
pyngrok==7.1.6 
protobuf==3.20.3
vidgear==0.3.2
